Add an uninstall script so the user can easily nuke
Included a simple uninstall script that deletes the files written to
.installed_files.txt.

USE WITH CAUTION.

This is not super-safe since, in principle, someone could add a random
file to this file. Since it runs without sudo, probably nothing
system-destabilizing will happen, but plenty of irritating things do not
qualify as system-destabilizing.

I will revisit this uninstall script and see if I can't figure out
something better.