Actually add `reinstall.sh`
For real this time, I've added the reinstall script. Same information
applies as is in commit e3d1f251f86fc1c0f585334d90b973fd8392a21d --- I
just flubbed it in such a minor way (and then pushed) that I'm not going
to bother trying to track downa  fix right now. If I really want to fix
the commit history, the damage will be as bad whenever I get around to
doing that as it is now.