Update setup.py for PyPI goodness
I've started using a real versioning system---I'll try to keep that up
from here on out. The metadata of the setup file is somewhat improved:
new long description, language and platform specification, etc.

This silly project will soon be uploaded to PyPI after a little bit of
administrative work on the repo.