Major refactor of `bar.py`; now a class
This commit marks a major refactor of the package primary functionality.
Now the progress bar is instantiated as a class with various attributes
we used to access at every iteration. Now we only have to set up the bar
on the first iteration, and it can jsut roll from there, which cuts down
on some of the negligible (though undesirable) overhead.