Bump version number (PyPI idiosyncracies)
Still sorting out my PyPI workflow. It is possible for the twine upload
to fail partway through, which will cause "filename already used" type
errors if attempting to reupload, which in turn requires a version bump,
which in turn requires a new commit-merge-push-build cycle.

Or maybe it doesn't, but this is what I've found so far. Anyone trolling
through this log who knows how to circumvent this problem: please do
share---assuming the advice isn't along the lines of "next time you do
that, don't."